Add files via upload
The idea of this project is to provide RSA encryption for embedded systems. The goal of this code is to allow embedded systems to encrypt their traffic with low cost taking advantage of their existing hardware level encryption. In order to do this nodes must be able to securely agree on an encryption key. The nodes will exchange this information via RSA and use the agreed upon key for hardware level encryption. The RSAP file provides implementation for encryption, and decryption. I am working to add key generation support. The primeP file provides generation of large random numbers and uses Fermat's Primality Test to generate prime numbers.